server/pyproject.toml
[tool.poetry]
name = "todo"
version = "0.1.0"
description = "Todo List API"
authors = ["Saskevich Aleksandr <your.email@example.com>"]
license = "MIT"
packages = [{include = "todo"}]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = ">=0.115.12,<0.116.0"
uvicorn = {extras = ["standard"], version = ">=0.34.0,<0.35.0"}
pydantic = {extras = ["email"], version = ">=2.11.3,<3.0.0"}
pydantic-settings = ">=2.8.1,<3.0.0"
sqlalchemy = {extras = ["asyncio"], version = ">=2.0.40,<3.0.0"}
asyncpg = ">=0.30.0,<0.31.0"
alembic = ">=1.15.2,<2.0.0"
fastapi-users = {extras = ["sqlalchemy"], version = ">=14.0.1,<15.0.0"}
gunicorn = ">=23.0.0,<24.0.0"
aiohttp = ">=3.11.16,<4.0.0"
orjson = "^3.10.16"
fastapi-users-db-sqlalchemy = "^7.0.0"
urllib3 = "^2.4.0"
aiobotocore = "2.14.0"
botocore = "1.35.7"
celery = "^5.5.2"
redis = "^6.0.0"
stripe = "^12.1.0"
google-cloud-storage = "^3.1.0"
google-auth-oauthlib = "^1.2.2"
google-auth = "^2.40.3"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.171.0"
pg8000 = "^1.31.4"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
